As mentioned previously, the mongoDB in a dockerized setup takes the name of the service as defined in , which is MongoDB shell version v4.0.10 connecting to: mongodb://127.0.0.1:27017/?gssapiServiceName=mongodb Implicit session: session { "id" : UUID("b23df365-f8ef-4cb7-a512-39473e806496") } MongoDB server version: 4.0.10 Server has startup warnings: 2019-06-25T17:25:42.078+0200 I STORAGE  [initandlisten] 2019-06-25T17:25:42.078+0200 I STORAGE  [initandlisten] ** WARNING: Using the XFS filesystem is strongly recommended with the WiredTiger storage engine 2019-06-25T17:25:42.078+0200 I STORAGE  [initandlisten] **          See http://dochub.mongodb.org/core/prodnotes-filesystem 2019-06-25T17:25:42.756+0200 I CONTROL  [initandlisten] 2019-06-25T17:25:42.756+0200 I CONTROL  [initandlisten] ** WARNING: Access control is not enabled for the database. 2019-06-25T17:25:42.756+0200 I CONTROL  [initandlisten] **          Read and write access to data and configuration is unrestricted. 2019-06-25T17:25:42.756+0200 I CONTROL  [initandlisten] --- Enable MongoDB's free cloud-based monitoring service, which will then receive and display metrics about your deployment (disk utilization, CPU, operation statistics, etc).
The monitoring data will be available on a MongoDB website with a unique URL accessible to you
and anyone you share the URL with. MongoDB may use this information to make product
improvements and to suggest MongoDB products and deployment options to you.

To enable free monitoring, run the following command: db.enableFreeMonitoring()
To permanently disable this reminder, run the following command: db.disableFreeMonitoring()
---

> [3G[J[3G^C
bye in our case; this will *not* work if you install MongoDB locally (you will need to revert to ), but if you do so, please do not commit that change; the version on the master branch should be fit for _dockerized_ deployment, and I would recommend you to work with Docker too.